+++ crates/objc2-exception-helper/CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 * Fixed the symbol name to include the correct SemVer version of the crate.
+* Fixed the ABI of `try_catch` to be `extern "C-unwind"`.
+* Clarified that `try_catch` does not catch Rust panics.
 
 
 ## 0.1.0 - 2024-06-02